avoid exposing internal types
