+++ test/cert.test.ts
@@ -1,8 +1,8 @@
 import { describe, afterEach, test, expect } from "vitest";
 import forge from "node-forge";
 import { resolve } from "pathe";
-import { Listener, TLSCertOptions } from "../src";
-import { _private } from "../src/cert";
+import { Listener } from "../src";
+import { TLSCertOptions, _private } from "../src/cert";
 
 // eslint-disable-next-line no-console
 // console.log = fn()
